requirements.txt
streamlit==1.29.0
requests==2.31.0
beautifulsoup4==4.12.2
together==1.1.3
python-dotenv==1.0.0